requirements.txt
chardet==3.0.4
Django==1.9
django-extensions==1.9.9
django-rest-framework-mongoengine==3.3.1
djangorestframework==3.3.3
mongoengine==0.9.0
pdfminer.six==20170720
pycryptodome==3.4.7
pymongo==2.7
six==1.11.0
typing==3.6.2
